reorg kv_engine.cpp
