+++ doc/user_doc.md
@@ -112,7 +112,7 @@ There is an anonymous global collection with KV-Pairs directly accessible via Ge
 
 Users can also create named collections.
 
-KVDK currently supports sorted named collections. Users can iterate forward or backward starting from an arbitrary point(at a key or between two keys) by an iterator. Elements can also be directly accessed via SGet, SSet, SDelete operations.
+KVDK currently supports sorted named collections. Users can iterate forward or backward starting from an arbitrary point(at a key or between two keys) by an iterator. Elements can also be directly accessed via SortedGet, SortedSet, SortedDelete operations.
 
 ## Reads and Writes in Anonymous Global Collection
 
@@ -175,9 +175,9 @@ int main()
 
 ## Reads and Writes in a Named Collection
 
-A KVDK instance provides SGet, SSet, SDelete methods to query/modify/delete sorted entries. 
+A KVDK instance provides SortedGet, SortedSet, SortedDelete methods to query/modify/delete sorted entries. 
 
-The following code performs a series of SGet, SSet and SDelete operations, which also initialize a named collection implicitly.
+The following code performs a series of SortedGet, SortedSet and SortedDelete operations, which also initialize a named collection implicitly.
 
 ```c++
 int main()
@@ -197,46 +197,46 @@ int main()
     // Insert key1-value1 into "my_collection_1".
     // Implicitly create a collection named "my_collection_1" in which
     // key1-value1 is stored.
-    status = engine->SSet(collection1, key1, value1);
+    status = engine->SortedSet(collection1, key1, value1);
     assert(status == kvdk::Status::Ok);
 
     // Get value1 by key1 in collection "my_collection_1"
-    status = engine->SGet(collection1, key1, &v);
+    status = engine->SortedGet(collection1, key1, &v);
     assert(status == kvdk::Status::Ok);
     assert(v == value1);
 
     // Insert key1-value2 into "my_collection_2".
     // Implicitly create a collection named "my_collection_2" in which
     // key1-value2 is stored.
-    status = engine->SSet(collection2, key1, value2);
+    status = engine->SortedSet(collection2, key1, value2);
     assert(status == kvdk::Status::Ok);
 
     // Get value2 by key1 in collection "my_collection_2"
-    status = engine->SGet(collection2, key1, &v);
+    status = engine->SortedGet(collection2, key1, &v);
     assert(status == kvdk::Status::Ok);
     assert(v == value2);
 
     // Get value1 by key1 in collection "my_collection_1"
     // key1-value2 is stored in "my_collection_2"
     // Thus key1-value1 stored in "my_collection_1" is unaffected by operation
-    // engine->SSet(collection2, key1, value2).
-    status = engine->SGet(collection1, key1, &v);
+    // engine->SortedSet(collection2, key1, value2).
+    status = engine->SortedGet(collection1, key1, &v);
     assert(status == kvdk::Status::Ok);
     assert(v == value1);
 
     // Insert key2-value2 into collection "my_collection_2"
     // Collection "my_collection_2" already exists and no implicit collection
     // creation occurs.
-    status = engine->SSet(collection2, key2, value2);
+    status = engine->SortedSet(collection2, key2, value2);
     assert(status == kvdk::Status::Ok);
 
     // Delete key1-value1 in collection "my_collection_1"
     // Although "my_collection_1" has no elements now, the collection itself is
     // not deleted though.
-    status = engine->SDelete(collection1, key1);
+    status = engine->SortedDelete(collection1, key1);
     assert(status == kvdk::Status::Ok);
 
-    printf("Successfully performed SGet, SSet, SDelete operations on named "
+    printf("Successfully performed SortedGet, SortedSet, SortedDelete operations on named "
            "collections.\n");
   }
 
@@ -275,7 +275,7 @@ int main()
     for (int i = 0; i < 10; ++i) {
       // Collection "my_sorted_collection" is implicitly created in first
       // iteration
-      status = engine->SSet(sorted_collection, kv_pairs[i].first,
+      status = engine->SortedSet(sorted_collection, kv_pairs[i].first,
                             kv_pairs[i].second);
       assert(status == kvdk::Status::Ok);
     }
