+++ engine/kv_engine.cpp
@@ -20,8 +20,6 @@
 #include "configs.hpp"
 #include "dram_allocator.hpp"
 #include "kvdk/engine.hpp"
-#include "sorted_collection/iterator.hpp"
-#include "sorted_collection/skiplist.hpp"
 #include "structures.hpp"
 #include "utils/sync_point.hpp"
 #include "utils/utils.hpp"
@@ -181,148 +179,6 @@ Status KVEngine::Init(const std::string& name, const Configs& configs) {
   return s;
 }
 
-Status KVEngine::CreateSortedCollection(
-    const StringView collection_name,
-    const SortedCollectionConfigs& s_configs) {
-  Status s = MaybeInitAccessThread();
-  defer(ReleaseAccessThread());
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  if (!CheckKeySize(collection_name)) {
-    return Status::InvalidDataSize;
-  }
-
-  // TODO jiayu: package sorted collection creation and destroy in Skiplist
-  // class
-  auto hint = hash_table_->GetHint(collection_name);
-  std::lock_guard<SpinMutex> lg(*hint.spin);
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-  TimeStampType new_ts = holder.Timestamp();
-  auto ret =
-      lookupKey<true>(collection_name, SortedHeader | SortedHeaderDelete);
-  if (ret.s == NotFound || ret.s == Outdated) {
-    DLRecord* existing_header =
-        ret.s == Outdated ? ret.entry.GetIndex().skiplist->HeaderRecord()
-                          : nullptr;
-    auto comparator = comparators_.GetComparator(s_configs.comparator_name);
-    if (comparator == nullptr) {
-      GlobalLogger.Error("Compare function %s is not registered\n",
-                         s_configs.comparator_name);
-      return Status::Abort;
-    }
-    CollectionIDType id = list_id_.fetch_add(1);
-    std::string value_str =
-        Skiplist::EncodeSortedCollectionValue(id, s_configs);
-    uint32_t request_size =
-        sizeof(DLRecord) + collection_name.size() + value_str.size();
-    SpaceEntry space_entry = pmem_allocator_->Allocate(request_size);
-    if (space_entry.size == 0) {
-      return Status::PmemOverflow;
-    }
-    // PMem level of skiplist is circular, so the next and prev pointers of
-    // header point to itself
-    DLRecord* pmem_record = DLRecord::PersistDLRecord(
-        pmem_allocator_->offset2addr(space_entry.offset), space_entry.size,
-        new_ts, SortedHeader, pmem_allocator_->addr2offset(existing_header),
-        space_entry.offset, space_entry.offset, collection_name, value_str);
-
-    auto skiplist = std::make_shared<Skiplist>(
-        pmem_record, string_view_2_string(collection_name), id, comparator,
-        pmem_allocator_, hash_table_, s_configs.index_with_hashtable);
-    addSkiplistToMap(skiplist);
-    hash_table_->Insert(hint, ret.entry_ptr, SortedHeader, skiplist.get(),
-                        PointerType::Skiplist);
-  } else {
-    // Todo (jiayu): handle expired skiplist
-    // Todo (jiayu): what if skiplist exists but comparator not match?
-    return ret.s;
-  }
-
-  return Status::Ok;
-}
-
-Status KVEngine::DestroySortedCollection(const StringView collection_name) {
-  auto s = MaybeInitAccessThread();
-  defer(ReleaseAccessThread());
-  if (s != Status::Ok) {
-    return s;
-  }
-destroy_impl : {
-  auto hint = hash_table_->GetHint(collection_name);
-  std::unique_lock<SpinMutex> ul(*hint.spin);
-  auto snapshot_holder = version_controller_.GetLocalSnapshotHolder();
-  auto new_ts = snapshot_holder.Timestamp();
-  auto ret = lookupKey<false>(collection_name,
-                              static_cast<uint16_t>(RecordType::SortedHeader));
-  if (ret.s == Status::Ok) {
-    Skiplist* skiplist = ret.entry.GetIndex().skiplist;
-    DLRecord* header = skiplist->HeaderRecord();
-    assert(header->entry.meta.type == SortedHeader);
-    StringView value = header->Value();
-    auto request_size =
-        sizeof(DLRecord) + collection_name.size() + value.size();
-    SpaceEntry space_entry = pmem_allocator_->Allocate(request_size);
-    if (space_entry.size == 0) {
-      return Status::PmemOverflow;
-    }
-    DLRecord* pmem_record = DLRecord::PersistDLRecord(
-        pmem_allocator_->offset2addr_checked(space_entry.offset),
-        space_entry.size, new_ts, SortedHeaderDelete,
-        pmem_allocator_->addr2offset_checked(header), header->prev,
-        header->next, collection_name, value);
-    if (!Skiplist::Replace(header, pmem_record, hint.spin,
-                           skiplist->HeaderNode(), pmem_allocator_.get(),
-                           hash_table_.get())) {
-      goto destroy_impl;
-    }
-    hash_table_->Insert(hint, ret.entry_ptr, SortedHeaderDelete, skiplist,
-                        PointerType::Skiplist);
-    ul.unlock();
-    delayFree(OldDeleteRecord(pmem_record, ret.entry_ptr,
-                              PointerType::HashEntry, new_ts, hint.spin));
-    delayFree(OldDataRecord{header, new_ts});
-  } else if (ret.s == Status::Outdated || ret.s == Status::NotFound) {
-    ret.s = Status::Ok;
-  }
-  return ret.s;
-}
-}
-
-Iterator* KVEngine::NewSortedIterator(const StringView collection,
-                                      Snapshot* snapshot) {
-  Skiplist* skiplist;
-  // find collection
-  auto res = lookupKey<false>(collection, SortedHeader);
-  if (res.s == Status::Ok) {
-    skiplist = res.entry_ptr->GetIndex().skiplist;
-  }
-
-  bool create_snapshot = snapshot == nullptr;
-  if (create_snapshot) {
-    snapshot = GetSnapshot(false);
-  }
-
-  return res.s == Status::Ok
-             ? new SortedIterator(skiplist, pmem_allocator_,
-                                  static_cast<SnapshotImpl*>(snapshot),
-                                  create_snapshot)
-             : nullptr;
-}
-
-void KVEngine::ReleaseSortedIterator(Iterator* sorted_iterator) {
-  if (sorted_iterator == nullptr) {
-    GlobalLogger.Info("pass a nullptr in KVEngine::ReleaseSortedIterator!\n");
-    return;
-  }
-  SortedIterator* iter = static_cast<SortedIterator*>(sorted_iterator);
-  if (iter->own_snapshot_) {
-    ReleaseSnapshot(iter->snapshot_);
-  }
-  delete iter;
-}
-
 Status KVEngine::RestoreData() {
   Status s = MaybeInitAccessThread();
   if (s != Status::Ok) {
@@ -422,19 +278,17 @@ Status KVEngine::RestoreData() {
     switch (data_entry_cached.meta.type) {
       case RecordType::SortedElem:
       case RecordType::SortedElemDelete: {
-        s = RestoreSkiplistRecord(
-            static_cast<DLRecord*>(recovering_pmem_record));
+        s = restoreSortedElem(static_cast<DLRecord*>(recovering_pmem_record));
         break;
       }
       case RecordType::SortedHeaderDelete:
       case RecordType::SortedHeader: {
-        s = RestoreSkiplistHeader(
-            static_cast<DLRecord*>(recovering_pmem_record));
+        s = restoreSortedHeader(static_cast<DLRecord*>(recovering_pmem_record));
         break;
       }
       case RecordType::StringDataRecord:
       case RecordType::StringDeleteRecord: {
-        s = RestoreStringRecord(
+        s = restoreStringRecord(
             static_cast<StringRecord*>(recovering_pmem_record),
             data_entry_cached);
         break;
@@ -532,69 +386,6 @@ bool KVEngine::ValidateRecord(void* data_record) {
       kvdk_assert(false, "Unsupported type in ValidateRecord()!");
       return false;
   }
-}
-
-Status KVEngine::RestoreSkiplistHeader(DLRecord* header_record) {
-  return sorted_rebuilder_->AddHeader(header_record);
-}
-
-Status KVEngine::RestoreStringRecord(StringRecord* pmem_record,
-                                     const DataEntry& cached_entry) {
-  assert(pmem_record->entry.meta.type & StringRecordType);
-  if (RecoverToCheckpoint() &&
-      cached_entry.meta.timestamp > persist_checkpoint_->CheckpointTS()) {
-    purgeAndFree(pmem_record);
-    return Status::Ok;
-  }
-  auto view = pmem_record->Key();
-  std::string key{view.data(), view.size()};
-  DataEntry existing_data_entry;
-  HashEntry hash_entry;
-  HashEntry* entry_ptr = nullptr;
-
-  auto hint = hash_table_->GetHint(key);
-  std::lock_guard<SpinMutex> lg(*hint.spin);
-  Status s =
-      hash_table_->SearchForWrite(hint, key, StringRecordType, &entry_ptr,
-                                  &hash_entry, &existing_data_entry);
-
-  if (s == Status::MemoryOverflow) {
-    return s;
-  }
-
-  bool found = s == Status::Ok;
-  if (found &&
-      existing_data_entry.meta.timestamp >= cached_entry.meta.timestamp) {
-    purgeAndFree(pmem_record);
-    return Status::Ok;
-  }
-
-  hash_table_->Insert(hint, entry_ptr, cached_entry.meta.type, pmem_record,
-                      PointerType::StringRecord);
-  if (found) {
-    purgeAndFree(hash_entry.GetIndex().ptr);
-  }
-
-  return Status::Ok;
-}
-
-bool KVEngine::CheckAndRepairDLRecord(DLRecord* record) {
-  uint64_t offset = pmem_allocator_->addr2offset(record);
-  DLRecord* prev = pmem_allocator_->offset2addr<DLRecord>(record->prev);
-  DLRecord* next = pmem_allocator_->offset2addr<DLRecord>(record->next);
-  if (prev->next != offset && next->prev != offset) {
-    return false;
-  }
-  // Repair un-finished write
-  if (next && next->prev != offset) {
-    next->prev = offset;
-    pmem_persist(&next->prev, 8);
-  }
-  return true;
-}
-
-Status KVEngine::RestoreSkiplistRecord(DLRecord* pmem_record) {
-  return sorted_rebuilder_->AddElement(pmem_record);
 }
 
 Status KVEngine::PersistOrRecoverImmutableConfigs() {
@@ -888,150 +679,6 @@ Status KVEngine::Recovery() {
   return Status::Ok;
 }
 
-Status KVEngine::Get(const StringView key, std::string* value) {
-  Status s = MaybeInitAccessThread();
-
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  if (!CheckKeySize(key)) {
-    return Status::InvalidDataSize;
-  }
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-  auto ret = lookupKey<false>(key, StringDataRecord | StringDeleteRecord);
-  if (ret.s == Status::Ok) {
-    StringRecord* string_record = ret.entry.GetIndex().string_record;
-    kvdk_assert(string_record->GetRecordType() == StringDataRecord,
-                "Got wrong data type in string get");
-    kvdk_assert(string_record->Validate(), "Corrupted data in string get");
-    value->assign(string_record->Value().data(), string_record->Value().size());
-    return Status::Ok;
-  } else {
-    return ret.s == Status::Outdated ? Status::NotFound : ret.s;
-  }
-}
-
-Status KVEngine::Delete(const StringView key) {
-  Status s = MaybeInitAccessThread();
-
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  if (!CheckKeySize(key)) {
-    return Status::InvalidDataSize;
-  }
-
-  return StringDeleteImpl(key);
-}
-
-Status KVEngine::SDeleteImpl(Skiplist* skiplist, const StringView& user_key) {
-  std::string collection_key(skiplist->InternalKey(user_key));
-  if (!CheckKeySize(collection_key)) {
-    return Status::InvalidDataSize;
-  }
-
-  auto hint = hash_table_->GetHint(collection_key);
-
-  while (1) {
-    std::unique_lock<SpinMutex> ul(*hint.spin);
-    TimeStampType new_ts = version_controller_.GetCurrentTimestamp();
-
-    auto ret = skiplist->Delete(user_key, hint, new_ts);
-    switch (ret.s) {
-      case Status::Fail:
-        continue;
-      case Status::PmemOverflow:
-        return ret.s;
-      case Status::Ok:
-        ul.unlock();
-        if (ret.write_record != nullptr) {
-          kvdk_assert(
-              ret.existing_record != nullptr &&
-                  ret.existing_record->entry.meta.type == SortedElem,
-              "Wrong existing record type while insert a delete reocrd for "
-              "sorted collection");
-          delayFree(OldDataRecord{ret.existing_record, new_ts});
-          if (ret.hash_entry_ptr != nullptr) {
-            // delete record indexed by hash table
-            delayFree(OldDeleteRecord(ret.write_record, ret.hash_entry_ptr,
-                                      PointerType::HashEntry, new_ts,
-                                      hint.spin));
-          } else if (ret.dram_node != nullptr) {
-            // no hash index, by a skiplist node points to delete record
-            delayFree(OldDeleteRecord(ret.write_record, ret.dram_node,
-                                      PointerType::SkiplistNode, new_ts,
-                                      hint.spin));
-          } else {
-            // delete record nor pointed by hash entry nor skiplist node
-            delayFree(OldDeleteRecord(ret.write_record, nullptr,
-                                      PointerType::Empty, new_ts, hint.spin));
-          }
-        }
-        return ret.s;
-      default:
-        std::abort();
-    }
-    break;
-  }
-  return Status::Ok;
-}
-
-Status KVEngine::SortedSetImpl(Skiplist* skiplist, const StringView& user_key,
-                               const StringView& value) {
-  std::string collection_key(skiplist->InternalKey(user_key));
-  if (!CheckKeySize(collection_key) || !CheckValueSize(value)) {
-    return Status::InvalidDataSize;
-  }
-
-  auto hint = hash_table_->GetHint(collection_key);
-  while (1) {
-    std::unique_lock<SpinMutex> ul(*hint.spin);
-    TimeStampType new_ts = version_controller_.GetCurrentTimestamp();
-    auto ret = skiplist->Set(user_key, value, hint, new_ts);
-    switch (ret.s) {
-      case Status::Fail:
-        continue;
-      case Status::PmemOverflow:
-        break;
-      case Status::Ok:
-        if (ret.existing_record &&
-            ret.existing_record->entry.meta.type == SortedElem) {
-          ul.unlock();
-          delayFree(OldDataRecord{ret.existing_record, new_ts});
-        }
-        break;
-      default:
-        std::abort();  // never shoud reach
-    }
-    return ret.s;
-  }
-  return Status::Ok;
-}
-
-Status KVEngine::SSet(const StringView collection, const StringView user_key,
-                      const StringView value) {
-  Status s = MaybeInitAccessThread();
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  auto snapshot_holder = version_controller_.GetLocalSnapshotHolder();
-
-  Skiplist* skiplist = nullptr;
-
-  auto ret = lookupKey<false>(collection, SortedHeader | SortedHeaderDelete);
-  if (ret.s != Status::Ok) {
-    return ret.s == Status::Outdated ? Status::NotFound : ret.s;
-  }
-
-  kvdk_assert(ret.entry.GetIndexType() == PointerType::Skiplist,
-              "pointer type of skiplist in hash entry should be skiplist");
-  skiplist = ret.entry.GetIndex().skiplist;
-  return SortedSetImpl(skiplist, user_key, value);
-}
-
 Status KVEngine::CheckConfigs(const Configs& configs) {
   auto is_2pown = [](uint64_t n) { return (n > 0) && (n & (n - 1)) == 0; };
 
@@ -1094,30 +741,6 @@ Status KVEngine::CheckConfigs(const Configs& configs) {
   }
 
   return Status::Ok;
-}
-
-Status KVEngine::SDelete(const StringView collection,
-                         const StringView user_key) {
-  Status s = MaybeInitAccessThread();
-  if (s != Status::Ok) {
-    return s;
-  }
-  // Hold current snapshot in this thread
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-
-  Skiplist* skiplist = nullptr;
-  auto ret = lookupKey<false>(collection, SortedHeader | SortedHeaderDelete);
-  // GlobalLogger.Debug("ret.s is %d\n", ret.s);
-  if (ret.s != Status::Ok) {
-    return (ret.s == Status::Outdated || ret.s == Status::NotFound) ? Status::Ok
-                                                                    : ret.s;
-  }
-
-  kvdk_assert(ret.entry.GetIndexType() == PointerType::Skiplist,
-              "pointer type of skiplist in hash entry should be skiplist");
-  skiplist = ret.entry.GetIndex().skiplist;
-
-  return SDeleteImpl(skiplist, user_key);
 }
 
 Status KVEngine::MaybeInitPendingBatchFile() {
@@ -1247,191 +870,6 @@ Status KVEngine::BatchWrite(const WriteBatch& write_batch) {
     }
   }
   return Status::Ok;
-}
-
-Status KVEngine::Modify(const StringView key, ModifyFunc modify_func,
-                        void* modify_args, const WriteOptions& write_options) {
-  int64_t base_time = TimeUtils::millisecond_time();
-  if (write_options.ttl_time <= 0 ||
-      !TimeUtils::CheckTTL(write_options.ttl_time, base_time)) {
-    return Status::InvalidArgument;
-  }
-
-  ExpireTimeType expired_time = write_options.ttl_time == kPersistTime
-                                    ? kPersistTime
-                                    : write_options.ttl_time + base_time;
-
-  Status s = MaybeInitAccessThread();
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  auto hint = hash_table_->GetHint(key);
-  std::unique_lock<SpinMutex> ul(*hint.spin);
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-  TimeStampType new_ts = holder.Timestamp();
-  auto ret = lookupKey<true>(key, static_cast<RecordType>(StringRecordType));
-
-  StringRecord* existing_record = nullptr;
-  std::string existing_value;
-  std::string new_value;
-  // push it into cleaner
-  if (ret.s == Status::Ok) {
-    existing_record = ret.entry.GetIndex().string_record;
-    existing_value.assign(existing_record->Value().data(),
-                          existing_record->Value().size());
-  } else if (ret.s == Status::Outdated) {
-    existing_record = ret.entry.GetIndex().string_record;
-  } else if (ret.s == Status::NotFound) {
-    // nothing todo
-  } else {
-    return ret.s;
-  }
-
-  auto modify_operation = modify_func(
-      ret.s == Status::Ok ? &existing_value : nullptr, &new_value, modify_args);
-  switch (modify_operation) {
-    case ModifyOperation::Write: {
-      if (!CheckValueSize(new_value)) {
-        return Status::InvalidDataSize;
-      }
-      SpaceEntry space_entry =
-          pmem_allocator_->Allocate(StringRecord::RecordSize(key, new_value));
-      if (space_entry.size == 0) {
-        return Status::PmemOverflow;
-      }
-
-      StringRecord* new_record =
-          pmem_allocator_->offset2addr_checked<StringRecord>(
-              space_entry.offset);
-      StringRecord::PersistStringRecord(
-          new_record, space_entry.size, new_ts, StringDataRecord,
-          existing_record == nullptr
-              ? kNullPMemOffset
-              : pmem_allocator_->addr2offset_checked(existing_record),
-          key, new_value, expired_time);
-      hash_table_->Insert(hint, ret.entry_ptr, StringDataRecord, new_record,
-                          PointerType::StringRecord);
-      if (ret.s == Status::Ok) {
-        ul.unlock();
-        delayFree(OldDataRecord{existing_record, new_ts});
-      }
-      break;
-    }
-    case ModifyOperation::Delete: {
-      if (ret.s == Status::Ok) {
-        SpaceEntry space_entry =
-            pmem_allocator_->Allocate(StringRecord::RecordSize(key, ""));
-        if (space_entry.size == 0) {
-          return Status::PmemOverflow;
-        }
-
-        void* pmem_ptr =
-            pmem_allocator_->offset2addr_checked(space_entry.offset);
-        StringRecord::PersistStringRecord(
-            pmem_ptr, space_entry.size, new_ts, StringDeleteRecord,
-            pmem_allocator_->addr2offset_checked(existing_record), key, "");
-        hash_table_->Insert(hint, ret.entry_ptr, StringDeleteRecord, pmem_ptr,
-                            PointerType::StringRecord);
-        ul.unlock();
-        delayFree(OldDataRecord{ret.entry.GetIndex().string_record, new_ts});
-        delayFree(OldDeleteRecord(pmem_ptr, ret.entry_ptr,
-                                  PointerType::HashEntry, new_ts, hint.spin));
-      }
-      break;
-    }
-    case ModifyOperation::Abort: {
-      return Status::Abort;
-    }
-    case ModifyOperation::Noop: {
-      return Status::Ok;
-    }
-  }
-
-  return Status::Ok;
-}
-
-Status KVEngine::StringBatchWriteImpl(const WriteBatch::KV& kv,
-                                      BatchWriteHint& batch_hint) {
-  DataEntry data_entry;
-  HashEntry hash_entry;
-  HashEntry* entry_ptr = nullptr;
-
-  {
-    auto& hash_hint = batch_hint.hash_hint;
-    // hash table for the hint should be alread locked, so we do not lock it
-    // here
-    Status s =
-        hash_table_->SearchForWrite(hash_hint, kv.key, StringRecordType,
-                                    &entry_ptr, &hash_entry, &data_entry);
-    if (s == Status::MemoryOverflow) {
-      return s;
-    }
-    batch_hint.hash_entry_ptr = entry_ptr;
-    bool found = s == Status::Ok;
-
-    // Deleting kv is not existing
-    if (kv.type == StringDeleteRecord && !found) {
-      batch_hint.space_not_used = true;
-      return Status::Ok;
-    }
-
-    kvdk_assert(!found || batch_hint.timestamp >= data_entry.meta.timestamp,
-                "ts of new data smaller than existing data in batch write");
-
-    void* block_base =
-        pmem_allocator_->offset2addr(batch_hint.allocated_space.offset);
-
-    TEST_SYNC_POINT(
-        "KVEngine::BatchWrite::StringBatchWriteImpl::Pesistent::Before");
-
-    StringRecord::PersistStringRecord(
-        block_base, batch_hint.allocated_space.size, batch_hint.timestamp,
-        static_cast<RecordType>(kv.type),
-        found ? pmem_allocator_->addr2offset_checked(
-                    hash_entry.GetIndex().string_record)
-              : kNullPMemOffset,
-        kv.key, kv.type == StringDataRecord ? kv.value : "");
-
-    hash_table_->Insert(hash_hint, entry_ptr, (RecordType)kv.type, block_base,
-                        PointerType::StringRecord);
-
-    if (found) {
-      if (kv.type == StringDeleteRecord) {
-        batch_hint.delete_record_to_free = block_base;
-      }
-      if (hash_entry.GetRecordType() == StringDataRecord) {
-        batch_hint.data_record_to_free = hash_entry.GetIndex().string_record;
-      }
-    }
-  }
-
-  return Status::Ok;
-}
-
-Status KVEngine::SGet(const StringView collection, const StringView user_key,
-                      std::string* value) {
-  Status s = MaybeInitAccessThread();
-
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  // Hold current snapshot in this thread
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-
-  Skiplist* skiplist = nullptr;
-  auto ret = lookupKey<false>(collection, SortedHeader | SortedHeaderDelete);
-  if (ret.s != Status::Ok) {
-    return ret.s == Status::Outdated ? Status::NotFound : ret.s;
-  }
-
-  kvdk_assert(ret.entry.GetIndexType() == PointerType::Skiplist,
-              "pointer type of skiplist in hash entry should be skiplist");
-  skiplist = ret.entry.GetIndex().skiplist;
-
-  assert(skiplist);
-  return skiplist->Get(user_key, value);
 }
 
 Status KVEngine::GetTTL(const StringView str, TTLType* ttl_time) {
@@ -1564,118 +1002,6 @@ start_expire : {
   return res.s;
 }
 }
-Status KVEngine::StringDeleteImpl(const StringView& key) {
-  auto hint = hash_table_->GetHint(key);
-  std::unique_lock<SpinMutex> ul(*hint.spin);
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-  TimeStampType new_ts = holder.Timestamp();
-
-  auto ret = lookupKey<false>(key, StringDeleteRecord | StringDataRecord);
-  if (ret.s == Status::Ok) {
-    // We only write delete record if key exist
-    auto request_size = key.size() + sizeof(StringRecord);
-    SpaceEntry space_entry = pmem_allocator_->Allocate(request_size);
-    if (space_entry.size == 0) {
-      return Status::PmemOverflow;
-    }
-
-    void* pmem_ptr = pmem_allocator_->offset2addr_checked(space_entry.offset);
-    StringRecord::PersistStringRecord(pmem_ptr, space_entry.size, new_ts,
-                                      StringDeleteRecord,
-                                      pmem_allocator_->addr2offset_checked(
-                                          ret.entry.GetIndex().string_record),
-                                      key, "");
-    hash_table_->Insert(hint, ret.entry_ptr, StringDeleteRecord, pmem_ptr,
-                        PointerType::StringRecord);
-    ul.unlock();
-    delayFree(OldDataRecord{ret.entry.GetIndex().string_record, new_ts});
-    // Free this delete record to recycle PMem and DRAM space
-    delayFree(OldDeleteRecord(pmem_ptr, ret.entry_ptr, PointerType::HashEntry,
-                              new_ts, hint.spin));
-  }
-
-  return (ret.s == Status::NotFound || ret.s == Status::Outdated) ? Status::Ok
-                                                                  : ret.s;
-}
-
-Status KVEngine::StringSetImpl(const StringView& key, const StringView& value,
-                               const WriteOptions& write_options) {
-  int64_t base_time = TimeUtils::millisecond_time();
-  if (write_options.ttl_time <= 0 ||
-      !TimeUtils::CheckTTL(write_options.ttl_time, base_time)) {
-    return Status::InvalidArgument;
-  }
-
-  ExpireTimeType expired_time =
-      TimeUtils::TTLToExpireTime(write_options.ttl_time, base_time);
-
-  KeyStatus entry_status =
-      expired_time != kPersistTime ? KeyStatus::Volatile : KeyStatus::Persist;
-
-  auto hint = hash_table_->GetHint(key);
-  TEST_SYNC_POINT("KVEngine::StringSetImpl::BeforeLock");
-  std::unique_lock<SpinMutex> ul(*hint.spin);
-  auto holder = version_controller_.GetLocalSnapshotHolder();
-  TimeStampType new_ts = holder.Timestamp();
-
-  // Lookup key in hashtable
-  auto ret = lookupKey<true>(key, StringDataRecord | StringDeleteRecord);
-  if (ret.s == Status::MemoryOverflow || ret.s == Status::WrongType) {
-    return ret.s;
-  }
-
-  kvdk_assert(ret.s == Status::NotFound || ret.s == Status::Ok ||
-                  ret.s == Status::Outdated,
-              "Wrong return status in lookupKey in StringSetImpl");
-  StringRecord* existing_record =
-      ret.s == Status::NotFound ? nullptr : ret.entry.GetIndex().string_record;
-  kvdk_assert(!existing_record || new_ts > existing_record->GetTimestamp(),
-              "existing record has newer timestamp or wrong return status in "
-              "string set");
-
-  // Persist key-value pair to PMem
-  uint32_t requested_size = value.size() + key.size() + sizeof(StringRecord);
-  SpaceEntry space_entry = pmem_allocator_->Allocate(requested_size);
-  if (space_entry.size == 0) {
-    return Status::PmemOverflow;
-  }
-  StringRecord* new_record =
-      pmem_allocator_->offset2addr_checked<StringRecord>(space_entry.offset);
-  StringRecord::PersistStringRecord(
-      new_record, space_entry.size, new_ts, StringDataRecord,
-      pmem_allocator_->addr2offset(existing_record), key, value, expired_time);
-
-  hash_table_->Insert(hint, ret.entry_ptr, StringDataRecord, new_record,
-                      PointerType::StringRecord, entry_status);
-  // Free existing record
-  bool need_free =
-      existing_record && ret.entry.GetRecordType() != StringDeleteRecord &&
-      !ret.entry.IsExpiredStatus() /*Check if expired_key already handled by
-                                       background cleaner*/
-      ;
-
-  if (need_free) {
-    ul.unlock();
-    delayFree(OldDataRecord{ret.entry.GetIndex().string_record, new_ts});
-  }
-
-  return Status::Ok;
-}
-
-Status KVEngine::Set(const StringView key, const StringView value,
-                     const WriteOptions& options) {
-  Status s = MaybeInitAccessThread();
-  if (s != Status::Ok) {
-    return s;
-  }
-
-  if (!CheckKeySize(key) || !CheckValueSize(value)) {
-    return Status::InvalidDataSize;
-  }
-
-  return StringSetImpl(key, value, options);
-}
-
 }  // namespace KVDK_NAMESPACE
 
 // lookupKey
