+++ engine/kv_engine_hash.cpp
@@ -1,3 +1,7 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ * Copyright(c) 2021 Intel Corporation
+ */
+
 #include "kv_engine.hpp"
 
 namespace KVDK_NAMESPACE {
