+++ tests/tests.cpp
@@ -143,18 +143,18 @@ class EngineBasicTest : public testing::Test {
     auto SortedSetFunc = [&](const std::string& collection,
                              const std::string& key,
                              const std::string& value) -> Status {
-      return engine->SSet(collection, key, value);
+      return engine->SortedSet(collection, key, value);
     };
 
     auto SortedGetFunc = [&](const std::string& collection,
                              const std::string& key,
                              std::string* value) -> Status {
-      return engine->SGet(collection, key, value);
+      return engine->SortedGet(collection, key, value);
     };
 
     auto SortedDeleteFunc = [&](const std::string& collection,
                                 const std::string& key) -> Status {
-      return engine->SDelete(collection, key);
+      return engine->SortedDelete(collection, key);
     };
 
     auto SortedDestroyFunc = [&](const std::string& collection) {
@@ -182,18 +182,18 @@ class EngineBasicTest : public testing::Test {
     auto SortedSetFunc = [&](const std::string& collection,
                              const std::string& key,
                              const std::string& value) -> Status {
-      return engine->SSet(collection, key, value);
+      return engine->SortedSet(collection, key, value);
     };
 
     auto SortedGetFunc = [&](const std::string& collection,
                              const std::string& key,
                              std::string* value) -> Status {
-      return engine->SGet(collection, key, value);
+      return engine->SortedGet(collection, key, value);
     };
 
     auto SortedDeleteFunc = [&](const std::string& collection,
                                 const std::string& key) -> Status {
-      return engine->SDelete(collection, key);
+      return engine->SortedDelete(collection, key);
     };
 
     auto SortedDestroyFunc = [&](const std::string& collection) {
@@ -384,14 +384,14 @@ TEST_F(EngineBasicTest, TestUniqueKey) {
     // Create
     ASSERT_EQ(engine->CreateSortedCollection(collection_name), ret_s);
     // Set
-    ASSERT_EQ(engine->SSet(collection_name, elem_key, new_val), ret_s);
+    ASSERT_EQ(engine->SortedSet(collection_name, elem_key, new_val), ret_s);
     // Get
-    ASSERT_EQ(engine->SGet(collection_name, elem_key, &got_val), ret_s);
+    ASSERT_EQ(engine->SortedGet(collection_name, elem_key, &got_val), ret_s);
     if (ret_s == Status::Ok) {
       ASSERT_EQ(got_val, new_val);
     }
     // Delete elem
-    ASSERT_EQ(engine->SDelete(collection_name, elem_key), ret_s);
+    ASSERT_EQ(engine->SortedDelete(collection_name, elem_key), ret_s);
   }
 
   // Test unordered
@@ -489,8 +489,8 @@ TEST_F(EngineBasicTest, TestBasicSnapshot) {
       WriteOptions write_options;
       ASSERT_EQ(engine->Set(key1, key1, write_options), Status::Ok);
       ASSERT_EQ(engine->Set(key2, key2, write_options), Status::Ok);
-      ASSERT_EQ(engine->SSet(sorted_collection, key1, key1), Status::Ok);
-      ASSERT_EQ(engine->SSet(sorted_collection, key2, key2), Status::Ok);
+      ASSERT_EQ(engine->SortedSet(sorted_collection, key1, key1), Status::Ok);
+      ASSERT_EQ(engine->SortedSet(sorted_collection, key2, key2), Status::Ok);
     }
     // Wait snapshot done
     set_finished_threads.fetch_add(1);
@@ -511,11 +511,11 @@ TEST_F(EngineBasicTest, TestBasicSnapshot) {
       ASSERT_EQ(engine->Set(key1, "updated " + key1), Status::Ok);
       ASSERT_EQ(engine->Delete(key1), Status::Ok);
       ASSERT_EQ(engine->Set(key3, key3), Status::Ok);
-      ASSERT_EQ(engine->SSet(sorted_collection, key1, "updated " + key1),
+      ASSERT_EQ(engine->SortedSet(sorted_collection, key1, "updated " + key1),
                 Status::Ok);
-      ASSERT_EQ(engine->SDelete(sorted_collection, key2), Status::Ok);
-      ASSERT_EQ(engine->SSet(sorted_collection, key3, key3), Status::Ok);
-      ASSERT_EQ(engine->SSet(sorted_collection_after_snapshot, key1, key1), Ok);
+      ASSERT_EQ(engine->SortedDelete(sorted_collection, key2), Status::Ok);
+      ASSERT_EQ(engine->SortedSet(sorted_collection, key3, key3), Status::Ok);
+      ASSERT_EQ(engine->SortedSet(sorted_collection_after_snapshot, key1, key1), Ok);
     }
   };
 
@@ -599,24 +599,24 @@ TEST_F(EngineBasicTest, TestBasicSnapshot) {
         ASSERT_EQ(got_v1, key1);
         ASSERT_EQ(got_v2, key2);
 
-        ASSERT_EQ(backup_engine->SGet(sorted_collection, key1, &got_v1),
+        ASSERT_EQ(backup_engine->SortedGet(sorted_collection, key1, &got_v1),
                   Status::Ok);
-        ASSERT_EQ(backup_engine->SGet(sorted_collection, key2, &got_v2),
+        ASSERT_EQ(backup_engine->SortedGet(sorted_collection, key2, &got_v2),
                   Status::Ok);
-        ASSERT_EQ(backup_engine->SGet(sorted_collection, key3, &got_v3),
+        ASSERT_EQ(backup_engine->SortedGet(sorted_collection, key3, &got_v3),
                   Status::NotFound);
         ASSERT_EQ(got_v1, key1);
         ASSERT_EQ(got_v2, key2);
-        ASSERT_EQ(backup_engine->SGet(sorted_collection_after_snapshot, key1,
+        ASSERT_EQ(backup_engine->SortedGet(sorted_collection_after_snapshot, key1,
                                       &got_v1),
                   Status::NotFound);
-        ASSERT_EQ(engine->SGet(sorted_collection, key1, &got_v1), Status::Ok);
-        ASSERT_EQ(engine->SGet(sorted_collection, key2, &got_v2), Status::Ok);
-        ASSERT_EQ(engine->SGet(sorted_collection, key3, &got_v3),
+        ASSERT_EQ(engine->SortedGet(sorted_collection, key1, &got_v1), Status::Ok);
+        ASSERT_EQ(engine->SortedGet(sorted_collection, key2, &got_v2), Status::Ok);
+        ASSERT_EQ(engine->SortedGet(sorted_collection, key3, &got_v3),
                   Status::NotFound);
         ASSERT_EQ(got_v1, key1);
         ASSERT_EQ(got_v2, key2);
-        ASSERT_EQ(engine->SGet(sorted_collection_after_snapshot, key1, &got_v1),
+        ASSERT_EQ(engine->SortedGet(sorted_collection_after_snapshot, key1, &got_v1),
                   Status::NotFound);
       }
     }
@@ -845,9 +845,9 @@ TEST_F(EngineBasicTest, TestSeek) {
   ASSERT_EQ(engine->CreateSortedCollection(collection), Status::Ok);
   uint64_t z = 0;
   auto zero_filled_str = uint64_to_string(z);
-  ASSERT_EQ(engine->SSet(collection, zero_filled_str, zero_filled_str),
+  ASSERT_EQ(engine->SortedSet(collection, zero_filled_str, zero_filled_str),
             Status::Ok);
-  ASSERT_EQ(engine->SGet(collection, zero_filled_str, &val), Status::Ok);
+  ASSERT_EQ(engine->SortedGet(collection, zero_filled_str, &val), Status::Ok);
   auto iter = engine->NewSortedIterator(collection);
   ASSERT_NE(iter, nullptr);
   iter->Seek(zero_filled_str);
@@ -856,11 +856,11 @@ TEST_F(EngineBasicTest, TestSeek) {
   // Test SeekToFirst
   collection.assign("col2");
   ASSERT_EQ(engine->CreateSortedCollection(collection), Status::Ok);
-  ASSERT_EQ(engine->SSet(collection, "foo", "bar"), Status::Ok);
-  ASSERT_EQ(engine->SGet(collection, "foo", &val), Status::Ok);
-  ASSERT_EQ(engine->SDelete(collection, "foo"), Status::Ok);
-  ASSERT_EQ(engine->SGet(collection, "foo", &val), Status::NotFound);
-  ASSERT_EQ(engine->SSet(collection, "foo2", "bar2"), Status::Ok);
+  ASSERT_EQ(engine->SortedSet(collection, "foo", "bar"), Status::Ok);
+  ASSERT_EQ(engine->SortedGet(collection, "foo", &val), Status::Ok);
+  ASSERT_EQ(engine->SortedDelete(collection, "foo"), Status::Ok);
+  ASSERT_EQ(engine->SortedGet(collection, "foo", &val), Status::NotFound);
+  ASSERT_EQ(engine->SortedSet(collection, "foo2", "bar2"), Status::Ok);
   engine->ReleaseSortedIterator(iter);
   iter = engine->NewSortedIterator(collection);
   ASSERT_NE(iter, nullptr);
@@ -978,19 +978,19 @@ TEST_F(EngineBasicTest, TestSortedRestore) {
           auto key = key_prefix + std::to_string(i);
           auto overall_val = std::to_string(i);
           auto t_val = std::to_string(i * 2);
-          ASSERT_EQ(engine->SSet(global_skiplist, key, overall_val),
+          ASSERT_EQ(engine->SortedSet(global_skiplist, key, overall_val),
                     Status::Ok);
-          ASSERT_EQ(engine->SSet(t_skiplist, key, t_val), Status::Ok);
-          ASSERT_EQ(engine->SGet(global_skiplist, key, &got_val), Status::Ok);
+          ASSERT_EQ(engine->SortedSet(t_skiplist, key, t_val), Status::Ok);
+          ASSERT_EQ(engine->SortedGet(global_skiplist, key, &got_val), Status::Ok);
           ASSERT_EQ(got_val, overall_val);
-          ASSERT_EQ(engine->SGet(t_skiplist, key, &got_val), Status::Ok);
+          ASSERT_EQ(engine->SortedGet(t_skiplist, key, &got_val), Status::Ok);
           ASSERT_EQ(got_val, t_val);
           if (i % 2 == 1) {
-            ASSERT_EQ(engine->SDelete(global_skiplist, key), Status::Ok);
-            ASSERT_EQ(engine->SDelete(t_skiplist, key), Status::Ok);
-            ASSERT_EQ(engine->SGet(global_skiplist, key, &got_val),
+            ASSERT_EQ(engine->SortedDelete(global_skiplist, key), Status::Ok);
+            ASSERT_EQ(engine->SortedDelete(t_skiplist, key), Status::Ok);
+            ASSERT_EQ(engine->SortedGet(global_skiplist, key, &got_val),
                       Status::NotFound);
-            ASSERT_EQ(engine->SGet(t_skiplist, key, &got_val),
+            ASSERT_EQ(engine->SortedGet(t_skiplist, key, &got_val),
                       Status::NotFound);
           }
         }
@@ -1016,14 +1016,14 @@ TEST_F(EngineBasicTest, TestSortedRestore) {
           std::string key(key_prefix + std::to_string(i));
           std::string overall_val(std::to_string(i));
           std::string t_val(std::to_string(i * 2));
-          Status s = engine->SGet(global_skiplist, key, &got_val);
+          Status s = engine->SortedGet(global_skiplist, key, &got_val);
           if (i % 2 == 1) {
             ASSERT_EQ(s, Status::NotFound);
           } else {
             ASSERT_EQ(s, Status::Ok);
             ASSERT_EQ(got_val, overall_val);
           }
-          s = engine->SGet(t_skiplist, key, &got_val);
+          s = engine->SortedGet(t_skiplist, key, &got_val);
           if (i % 2 == 1) {
             ASSERT_EQ(s, Status::NotFound);
           } else {
@@ -1137,19 +1137,19 @@ TEST_F(EngineBasicTest, TestMultiThreadSortedRestore) {
 
       auto key = key_prefix + std::to_string(i);
       auto average_val = std::to_string(i);
-      ASSERT_EQ(engine->SSet(average_skiplist, key, average_val), Status::Ok);
-      ASSERT_EQ(engine->SGet(average_skiplist, key, &got_val), Status::Ok);
+      ASSERT_EQ(engine->SortedSet(average_skiplist, key, average_val), Status::Ok);
+      ASSERT_EQ(engine->SortedGet(average_skiplist, key, &got_val), Status::Ok);
       ASSERT_EQ(got_val, average_val);
       auto r_val = std::to_string(i * 2);
-      ASSERT_EQ(engine->SSet(r_skiplist, key, r_val), Status::Ok);
-      ASSERT_EQ(engine->SGet(r_skiplist, key, &got_val), Status::Ok);
+      ASSERT_EQ(engine->SortedSet(r_skiplist, key, r_val), Status::Ok);
+      ASSERT_EQ(engine->SortedGet(r_skiplist, key, &got_val), Status::Ok);
       ASSERT_EQ(got_val, r_val);
       if ((rand() % i) == 0) {
-        ASSERT_EQ(engine->SDelete(average_skiplist, key), Status::Ok);
-        ASSERT_EQ(engine->SDelete(r_skiplist, key), Status::Ok);
-        ASSERT_EQ(engine->SGet(average_skiplist, key, &got_val),
+        ASSERT_EQ(engine->SortedDelete(average_skiplist, key), Status::Ok);
+        ASSERT_EQ(engine->SortedDelete(r_skiplist, key), Status::Ok);
+        ASSERT_EQ(engine->SortedGet(average_skiplist, key, &got_val),
                   Status::NotFound);
-        ASSERT_EQ(engine->SGet(r_skiplist, key, &got_val), Status::NotFound);
+        ASSERT_EQ(engine->SortedGet(r_skiplist, key, &got_val), Status::NotFound);
       }
     }
   };
@@ -1526,7 +1526,7 @@ TEST_F(EngineBasicTest, TestSortedHotspot) {
   ASSERT_EQ(engine->CreateSortedCollection(collection_name), Status::Ok);
 
   for (const std::string& key : keys) {
-    ASSERT_EQ(engine->SSet(collection_name, key, val1), Status::Ok);
+    ASSERT_EQ(engine->SortedSet(collection_name, key, val1), Status::Ok);
     engine->ReleaseAccessThread();
 
     auto EvenWriteOddRead = [&](uint32_t id) {
@@ -1534,14 +1534,14 @@ TEST_F(EngineBasicTest, TestSortedHotspot) {
         if (id % 2 == 0) {
           // Even Write
           if (id % 4 == 0) {
-            ASSERT_EQ(engine->SSet(collection_name, key, val1), Status::Ok);
+            ASSERT_EQ(engine->SortedSet(collection_name, key, val1), Status::Ok);
           } else {
-            ASSERT_EQ(engine->SSet(collection_name, key, val2), Status::Ok);
+            ASSERT_EQ(engine->SortedSet(collection_name, key, val2), Status::Ok);
           }
         } else {
           // Odd Read
           std::string got_val;
-          ASSERT_EQ(engine->SGet(collection_name, key, &got_val), Status::Ok);
+          ASSERT_EQ(engine->SortedGet(collection_name, key, &got_val), Status::Ok);
           bool match = false;
           match = match || (got_val == val1);
           match = match || (got_val == val2);
@@ -1627,7 +1627,7 @@ TEST_F(EngineBasicTest, TestSortedCustomCompareFunction) {
   for (size_t i = 0; i < collections.size(); ++i) {
     auto Write = [&](size_t) {
       for (size_t j = 0; j < count; j++) {
-        ASSERT_EQ(engine->SSet(collections[i], key_values[j].first,
+        ASSERT_EQ(engine->SortedSet(collections[i], key_values[j].first,
                                key_values[j].second),
                   Status::Ok);
       }
@@ -1687,7 +1687,7 @@ TEST_F(EngineBasicTest, TestHashTableIterator) {
       ASSERT_EQ(engine->Set("stringkey" + std::to_string(id), "stringval"),
                 Status::Ok);
     } else {
-      ASSERT_EQ(engine->SSet(collection_name, "sortedkey" + std::to_string(id),
+      ASSERT_EQ(engine->SortedSet(collection_name, "sortedkey" + std::to_string(id),
                              "sortedval"),
                 Status::Ok);
     }
@@ -1793,12 +1793,12 @@ TEST_F(EngineBasicTest, TestExpireAPI) {
   {
     Status s = engine->CreateSortedCollection(sorted_collection);
     ASSERT_EQ(s, Status::Ok);
-    ASSERT_EQ(engine->SSet(sorted_collection, "sorted" + key, "sorted" + val),
+    ASSERT_EQ(engine->SortedSet(sorted_collection, "sorted" + key, "sorted" + val),
               Status::Ok);
     // Set expired time for collection
     ASSERT_EQ(engine->Expire(sorted_collection, max_ttl_time),
               Status::InvalidArgument);
-    ASSERT_EQ(engine->SSet(sorted_collection, "sorted2" + key, "sorted2" + val),
+    ASSERT_EQ(engine->SortedSet(sorted_collection, "sorted2" + key, "sorted2" + val),
               Status::Ok);
     ASSERT_EQ(engine->GetTTL(sorted_collection, &ttl_time), Status::Ok);
     // check sorted_collection is persist;
@@ -1806,7 +1806,7 @@ TEST_F(EngineBasicTest, TestExpireAPI) {
     // reset expired time for collection
     ASSERT_EQ(engine->Expire(sorted_collection, 2), Status::Ok);
     sleep(2);
-    ASSERT_EQ(engine->SGet(sorted_collection, "sorted" + key, &got_val),
+    ASSERT_EQ(engine->SortedGet(sorted_collection, "sorted" + key, &got_val),
               Status::NotFound);
     ASSERT_EQ(engine->GetTTL(sorted_collection, &ttl_time), Status::NotFound);
     ASSERT_EQ(ttl_time, kInvalidTTL);
@@ -2131,7 +2131,7 @@ TEST_F(EngineBasicTest, TestSortedRecoverySyncPointCaseOne) {
         }
       });
   SyncPoint::GetInstance()->SetCallBack(
-      "Test::SSet::Update::Finish", [&](void*) { update_num.fetch_add(1); });
+      "Test::SortedSet::Update::Finish", [&](void*) { update_num.fetch_add(1); });
   SyncPoint::GetInstance()->EnableProcessing();
 
   std::string collection_name = "SortedRecoverySyncPoint";
@@ -2141,12 +2141,12 @@ TEST_F(EngineBasicTest, TestSortedRecoverySyncPointCaseOne) {
   try {
     for (int i = 0; i < cnt; ++i) {
       if (i % 2 == 0) {
-        engine->SSet(collection_name, "key" + std::to_string(i),
+        engine->SortedSet(collection_name, "key" + std::to_string(i),
                      "val" + std::to_string(i));
       } else {
         std::string new_val = "val*" + std::to_string(i);
-        engine->SSet(collection_name, "key" + std::to_string(i - 1), new_val);
-        TEST_SYNC_POINT("Test::SSet::Update::Finish");
+        engine->SortedSet(collection_name, "key" + std::to_string(i - 1), new_val);
+        TEST_SYNC_POINT("Test::SortedSet::Update::Finish");
       }
     }
   } catch (...) {
@@ -2157,7 +2157,7 @@ TEST_F(EngineBasicTest, TestSortedRecoverySyncPointCaseOne) {
     for (int i = 0; i < cnt; ++i) {
       std::string key = "key" + std::to_string(i);
       std::string got_val;
-      Status s = engine->SGet(collection_name, key, &got_val);
+      Status s = engine->SortedGet(collection_name, key, &got_val);
       if (i % 2 != 0) {
         ASSERT_EQ(s, Status::NotFound);
       } else {
@@ -2192,7 +2192,7 @@ TEST_F(EngineBasicTest, TestSortedRecoverySyncPointCaseTwo) {
   std::atomic<bool> first_visited{false};
   SyncPoint::GetInstance()->DisableProcessing();
   SyncPoint::GetInstance()->Reset();
-  // only throw when the first call `SDelete`
+  // only throw when the first call `SortedDelete`
   SyncPoint::GetInstance()->SetCallBack(
       "KVEngine::Skiplist::Delete::PersistNext'sPrev::After", [&](void*) {
         if (!first_visited.load()) {
@@ -2208,10 +2208,10 @@ TEST_F(EngineBasicTest, TestSortedRecoverySyncPointCaseTwo) {
 
   std::vector<std::string> keylists{"C", "A", "D"};
   try {
-    engine->SSet(collection_name, keylists[0], "val" + keylists[0]);
-    engine->SSet(collection_name, keylists[1], "val" + keylists[1]);
-    engine->SSet(collection_name, keylists[2], "val" + keylists[2]);
-    engine->SDelete(collection_name, keylists[0]);
+    engine->SortedSet(collection_name, keylists[0], "val" + keylists[0]);
+    engine->SortedSet(collection_name, keylists[1], "val" + keylists[1]);
+    engine->SortedSet(collection_name, keylists[2], "val" + keylists[2]);
+    engine->SortedDelete(collection_name, keylists[0]);
   } catch (...) {
     delete engine;
     // reopen engine
@@ -2236,14 +2236,14 @@ TEST_F(EngineBasicTest, TestSortedRecoverySyncPointCaseTwo) {
     ASSERT_EQ(forward_num, backward_num);
 
     std::string got_val;
-    ASSERT_EQ(engine->SGet(collection_name, keylists[0], &got_val), Status::Ok);
-    ASSERT_EQ(engine->SGet(collection_name, keylists[2], &got_val), Status::Ok);
+    ASSERT_EQ(engine->SortedGet(collection_name, keylists[0], &got_val), Status::Ok);
+    ASSERT_EQ(engine->SortedGet(collection_name, keylists[2], &got_val), Status::Ok);
     ASSERT_EQ(got_val, "val" + keylists[2]);
-    ASSERT_EQ(engine->SGet(collection_name, keylists[1], &got_val), Status::Ok);
+    ASSERT_EQ(engine->SortedGet(collection_name, keylists[1], &got_val), Status::Ok);
     ASSERT_EQ(got_val, "val" + keylists[1]);
 
     // Again delete "C".
-    ASSERT_EQ(engine->SDelete(collection_name, keylists[0]), Status::Ok);
+    ASSERT_EQ(engine->SortedDelete(collection_name, keylists[0]), Status::Ok);
   }
 }
 
@@ -2260,8 +2260,8 @@ TEST_F(EngineBasicTest, TestSortedSyncPoint) {
   std::string collection_name = "skiplist";
   ASSERT_EQ(engine->CreateSortedCollection(collection_name), Status::Ok);
 
-  engine->SSet(collection_name, "key0", "val0");
-  engine->SSet(collection_name, "key2", "val2");
+  engine->SortedSet(collection_name, "key0", "val0");
+  engine->SortedSet(collection_name, "key2", "val2");
 
   std::atomic<bool> first_record(false);
   SyncPoint::GetInstance()->DisableProcessing();
@@ -2272,9 +2272,9 @@ TEST_F(EngineBasicTest, TestSortedSyncPoint) {
 
   // insert
   ths.emplace_back(std::thread([&]() {
-    engine->SSet(collection_name, "key1", "val1");
+    engine->SortedSet(collection_name, "key1", "val1");
     std::string got_val;
-    ASSERT_EQ(engine->SGet(collection_name, "key1", &got_val), Status::Ok);
+    ASSERT_EQ(engine->SortedGet(collection_name, "key1", &got_val), Status::Ok);
   }));
 
   // Iter
@@ -2420,7 +2420,7 @@ TEST_F(EngineBasicTest, TestBackGroundCleaner) {
         s_configs.index_with_hashtable = index_with_hashtable;
         ASSERT_EQ(engine->CreateSortedCollection(sorted_collection, s_configs),
                   Status::Ok);
-        ASSERT_EQ(engine->SSet(sorted_collection, key, val), Status::Ok);
+        ASSERT_EQ(engine->SortedSet(sorted_collection, key, val), Status::Ok);
         bool set_expire = fast_random_64() % 2 == 0;
         if (set_expire) {
           ASSERT_EQ(engine->Expire(sorted_collection, 1), Status::Ok);
@@ -2441,7 +2441,7 @@ TEST_F(EngineBasicTest, TestBackGroundCleaner) {
         Status s = engine->GetTTL(sorted_collection, &ttl_time);
         if (s == Status::Ok) {
           if (ttl_time = kPersistTime) {
-            ASSERT_EQ(engine->SGet(sorted_collection, key, &got_val),
+            ASSERT_EQ(engine->SortedGet(sorted_collection, key, &got_val),
                       Status::Ok);
             ASSERT_EQ(got_val, val);
           } else {
@@ -2450,7 +2450,7 @@ TEST_F(EngineBasicTest, TestBackGroundCleaner) {
         } else {
           ASSERT_EQ(s, Status::NotFound);
           ASSERT_EQ(ttl_time, kInvalidTTL);
-          ASSERT_EQ(engine->SGet(sorted_collection, key, &got_val),
+          ASSERT_EQ(engine->SortedGet(sorted_collection, key, &got_val),
                     Status::NotFound);
         }
       }
